ETS - FroQ/pyproject.toml
[project]
name = "ets-translator"
version = "0.1.0"
description = "End-to-end translation project: training and inference"
readme = "README.md"
requires-python = ">=3.9"
authors = [
  { name = "FroQ", email = "dev@example.com" }
]
dependencies = [
  "pydantic>=2.7.0",
  "pyyaml>=6.0.1",
  "transformers>=4.43.0",
  "datasets>=2.20.0",
  "accelerate>=0.33.0",
  "sacrebleu>=2.4.2",
  "sentencepiece>=0.2.0",
  "evaluate>=0.4.2",
  "rich>=13.7.1",
  "typer>=0.12.0",
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.1",
  "torch>=2.3.0; platform_system != 'Darwin' or platform_machine != 'arm64'",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.5.0",
  "mypy>=1.10.0"
]

[project.scripts]
ets-translate = "ets_translator.inference.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]


